requirements.txt
flask
flask_restful
google.cloud.logging
google-cloud-pubsub
json